layers/utilities/requirements.txt
aws-cdk-lib==2.47.0
boto3==1.24.19
boto3-type-annotations==0.3.1
botocore==1.27.19
PyYAML==6.0